Move out transformation logic (#62)
The transformation logic has been moved to the importer (voyager). This is to enable fall forward in live migration (i.e. now, after this change, the exporter will export data in a target-agnostic format, and it is the importer's job to transform the data based on the target db type. 

Ref: https://github.com/yugabyte/yb-voyager/pull/946