track queue segment fsynced size in meta db (#73)
we  perform the following in quick succcession at the end of a batch in debezium:
- fsync
- store file size of segment to metadata db (on recovery we truncate file after this offset)
- commit batch to debezium

This greatly reduces the possibility that we would deliver an event more-than-once (process has to fail after storing file size and before committing batch. But more importantly, this ensures that the importer does not read un-fsynced data.

Note: however, we also fsync on rotation of queue segment, which increases the probability of more-than-once.. Need to consider preventing rotation in between a batch.