[yugabyte/yugabyte-db#27177] Introduce YSQL major upgrade config property (#185)
### Problem

With Yugbayte-DB's ysql upgrade to PG 15, the catalog tables are deleted
and recreated during the upgrade process. As a result the walsender
fails when it tries to read the catalog tables by setting `yb_read_time`
pointing to PG 11 era.

### Solution

With this PR we are introducing a new config property called
`ysql.major.upgrade`. The default value of this property is false. Any
user experiencing an error due to failure in walsender while reading
catalog table can set this property to true. When set, the connector
task will set the the `skip_yb_read_time_in_walsender` GUC in the
walsender session, whereby the catalog reads will be performed as of the
current time. Once the restart time crosses the update time, this config
property should be reset to false.

### Test Plan
Tested the changes manually by orchestrating the upgrade process
manually on yugabyted cluster and using custom docker image for the
connector.