[#28533] Make yb.load.balance.connections connector config a string instead of boolean (#184)
### Problem

Currently, the yb.load.balance.connections config is a boolean and so
only accepts either true or false. However, load balancing property of
smart drivers accepts additional values such as `any`, `only-primary`,
`only-rr`, `prefer-primary` and `prefer-rr`. Currently, the value `any`
is alias to yb.load.balance.connections' value `true`.

### Solution

To support these additional values of load balancing property, this PR
converts yb.load.balance.connections to string with default value being
**`only-primary`**. PR also changes the JDBC smart driver version to
`42.7.3-yb-4` since load-balance property (including advanced options
like `any`, `only-primary`, `only-rr`, etc.) is supported in version
`42.7.3-yb-1` and later.

### Test Plan

To validate the change, a YugabyteDB cluster was deployed and connected
to a custom docker image of the connector. The connection was
established to the correct node as per the value of the
yb.load.balance.connections property provided in connector creation
request (or defaulted to `only-primary` when no value of it was
provided).