[yugabyte/yugabyte-db#29499] Add support for publishing the connector jar to YB s3 maven (#187)
This PR adds dependencies required to publish the yugabytedb connector
to Yugabyte's s3 hosted maven repository.

#### List of changes made in this PR
1. Bumped up the version in all the pom files to the appropriate latest
value.
2. Added plugins necessary to publish the jar in correct format to
Yugabyte's s3 hosted maven repository.
3. Suppressed / removed the plugins inherited from the upstream code
which are used for pushing the jar to maven central repository. This was
needed because otherwise `mvn deploy` tries to publish to maven central
and fails, thus skipping publishing to Yugabyte's s3 hosted maven
 
#### Steps to publish a connector jar to s3 maven repository
1. Change the revision property in the root pom file to appropriate
value.
  2. Compile the jar.
4. Update the s3 credentials either in the env variables or in
`~/.s2/credentials` file. Set AWS_REGION=us-east-1, this is needed
because the s3 bucket is in us-east-1 region.
5. From inside project root directory run `mvn clean deploy -Ps3-deploy
-Dquick -pl debezium-connector-postgres -B`. This will publish a zip
file, a jar and a pom file to Yugabyte's s3 hosted maven repository. The
list and format of artefacts is kept the same as seen with
`debezium-connector-postgres` on maven central repository.