[DBZ-PGYB][yugabyte/yugabyte-db#29660] Upgrade Java version to Java 17 in YB connector (#190)
## Problem 
The current YB connector is based on Java 11. This PR upgrades the Java
version and necessary dependencies to Java 17-compatible versions.
The upstream Debezium project upgraded to Java 17 in version 3.0. While
we are currently on version 2.5.2, upgrading to Java 17 in this version
is safe and beneficial because it simplifies cherry-picking from
upstream patches from upstream (which uses Java 17) will not require
Java version modifications.


## Solution
To upgrade the Java version we will do the following:

- Update `pom.xml` and change the min. jdk version and maven release
version
- Update `debezium-schema-generator/pom.xml` and change
maven-plugin-plugin from 3.6.0 to 3.9.0. Required because 3.6.0 uses ASM
7.x which cannot read Java 17 bytecode.
- Update `Dockerfile` - Install Java 17 runtime

## Test Plan
Manual testing + Few existing